Changes With Review Comments
There are quit some of the file name and folder structure changes.

Fixes For Review Comments for  following items.
------------------------------------------------------
1. There is no plugin in the POM for running a Jetty or Tomcat server to
host the app. No support for running embedded Jetty or Tomcat has been
implemented, therefore the app requires an external Jetty or Tomcat
server

2.  The app also requires an external MySQL server and the schema is not
supplied with the app

3. The java.util.logging FileHandler path is hard coded to a Windows
path that does not exist on OS X.

4. Integration Test and Unit Test For Service and Persistence Layer